SA-1957: Group by table
